pyproject.toml
[project]
name = "titanoboa-favorites"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "titanoboa>=0.2.5",
]
